fix: cantidades decimales (kg, litros) se truncaban a enteros
FIX 18A - models/schema.py:
- quantity: INTEGER → REAL DEFAULT 1.0
- Migración automática: recrear transaction_items si quantity es INTEGER

FIX 18B - app.py:
- int(item_quants[i]) → float(item_quants[i]) en 4 sitios
- scan_save (2 sitios) + new_expense_manual (2 sitios)

FIX 18C - templates:
- scan/edit.html: min="1" → min="0.001" step="0.001"
- expenses/new.html: min="1" → min="0.001" step="0.001"
- expenses/detail.html: mostrar decimales solo cuando sea necesario

Causa: 1.412 kg de plátanos se guardaba como 1 (entero),
haciendo que items_sum no cuadrara con el total.

+++ app.py
@@ -410,9 +410,9 @@ def scan_save():
         except (ValueError, IndexError):
             price = 0
         try:
-            qty = int(item_quants[i])
+            qty = float(item_quants[i])
         except (ValueError, IndexError):
-            qty = 1
+            qty = 1.0
         if desc and price > 0:
             items_sum += price * qty
             c.execute("""
@@ -428,9 +428,9 @@ def scan_save():
         except (ValueError, IndexError):
             price = 0
         try:
-            qty = int(item_quants[i])
+            qty = float(item_quants[i])
         except (ValueError, IndexError):
-            qty = 1
+            qty = 1.0
         if desc and price > 0:
             # precio del OCR = precio UNITARIO (no total)
             # Para gasolina: precio = €/litro, cantidad = litros
@@ -1071,9 +1071,9 @@ def new_expense_manual():
             except (ValueError, IndexError):
                 price = 0
             try:
-                qty = int(item_quants[i])
+                qty = float(item_quants[i])
             except (ValueError, IndexError):
-                qty = 1
+                qty = 1.0
             if desc and price > 0:
                 c.execute("""
                     INSERT INTO transaction_items (transaction_id, description, quantity, unit_price, category_id)
@@ -1088,9 +1088,9 @@ def new_expense_manual():
             except (ValueError, IndexError):
                 price = 0
             try:
-                qty = int(item_quants[i])
+                qty = float(item_quants[i])
             except (ValueError, IndexError):
-                qty = 1
+                qty = 1.0
             if desc and price > 0:
                 # precio del form = precio UNITARIO (no total)
                 unit_price = price
